entregables/entregable_2/requirements.txt

yfinance
sqlalchemy-redshift
redshift_connector
python-dotenv
psycopg2-binary
pandas